src/test/java/com/gsg/event/domain/model/event/BirthdayEventTest.java
package com.gsg.event.domain.model.event;

import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
import com.amazonaws.services.dynamodbv2.model.CreateTableRequest;
import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
import com.gsg.event.config.config.SampleBean;
import com.gsg.kernel.usermgmt.UserId;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.web.WebAppConfiguration;


import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
@WebAppConfiguration
@ActiveProfiles("local")
@TestPropertySource(properties = {
        "amazon.dynamodb.endpoint=http://localhost:8000/",
        "amazon.aws.accesskey=test1",
        "amazon.aws.secretkey=test231" })
public class BirthdayEventTest
{
    private DynamoDBMapper dynamoDBMapper;

    @Autowired
    private AmazonDynamoDB amazonDynamoDB;

    @Autowired
    EventRepository repository;

    @Autowired
    SampleBean sb;

    private static final String EXPECTED_COST = "20";
    private static final String EXPECTED_PRICE = "50";

    @BeforeEach
    public void setup() throws Exception {
        sb.printSomething();

        dynamoDBMapper = new DynamoDBMapper(amazonDynamoDB);

        CreateTableRequest tableRequest = dynamoDBMapper
                .generateCreateTableRequest(BirthdayEvent.class);
        tableRequest.setProvisionedThroughput(
                new ProvisionedThroughput(1L, 1L));
        amazonDynamoDB.createTable(tableRequest);

        dynamoDBMapper.batchDelete(
                (List<BirthdayEvent>)repository.findAll());

    }

    @Test
    public void givenItemWithExpectedCost_whenRunFindAll_thenItemIsFound() {
//        assertThat(true);
        BirthdayEvent event = new BirthdayEvent(UserId.getNextId(), "Dhaval");

        repository.save(event);




        assertThat(result.size()).isGreaterThan(0);
        assertThat(result.get(0).getUUID()).isEqualTo(event.getUUID());
    }

}
